added the ability of adding all files at once
