+++ src/cli/exec.rs
@@ -1,6 +1,7 @@
 use crate::{
     cli::commands,
     ops::{add::add_rit, commit::commit_rit, init::init_rit, status::status_rit},
+    utils::ioutils::get_all_paths,
 };
 use std::path::PathBuf;
 
@@ -13,13 +14,20 @@ pub fn exec_cli() {
             if let Some(a) = sub_matches.get_one::<bool>("all") {
                 add_all = *a
             }
-            let paths = sub_matches
-                .get_many::<PathBuf>("PATH")
-                .into_iter()
-                .flatten()
-                .collect::<Vec<_>>();
+            let paths;
+            match add_all {
+                true => paths = get_all_paths(),
+                false => {
+                    paths = sub_matches
+                        .get_many::<PathBuf>("PATH")
+                        .into_iter()
+                        .flatten()
+                        .cloned()
+                        .collect::<Vec<_>>();
+                }
+            }
 
-            match add_rit(paths, add_all) {
+            match add_rit(paths) {
                 Ok(_) => println!("files added"),
                 Err(e) => eprintln!("{}", e.to_string()),
             };
